+++ approach_person/package.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>hand</name>
+  <name>approach_person</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
-  <maintainer email="kohei@todo.todo">kohei</maintainer>
+  <maintainer email="c1005073@planet.kanazawa-it.ac.jp">yusukepad</maintainer>
   <license>TODO: License declaration</license>
 
   <test_depend>ament_copyright</test_depend>
