+++ approach_person/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/approach_person
+[install]
+install_scripts=$base/lib/approach_person
