--- detect_pose/detect_pose/detect_pose_node.py
@@ -1,6 +1,0 @@
-def main():
-    print('Hi from detect_pose.')
-
-
-if __name__ == '__main__':
-    main()
