+++ grasp_bag/setup.py
@@ -27,12 +27,7 @@ setup(
             'scan_data_sensing_node = grasp_bag.scan_data_sensing_node:main',
             'bag_localization_node = grasp_bag.bag_localization_node:main',
             'grasp_bag_server = grasp_bag.grasp_bag_server:main',
-            'test_node = grasp_bag.test_node:main',
-            'happy_mini_master = grasp_bag.happy_mini_master:main',
-            'happy_mini_masterf = grasp_bag.happy_mini_masterf:main',
-            'happymini_sekkin = grasp_bag.happymini_sekkin:main',
-            'happy_hand_pose = grasp_bag.happy_hand_pose:detect_hand_pose',
-            'tamesi = grasp_bag.tamesi:main'
+            'test_node = grasp_bag.test_node:main'
         ],
     },
 )
