--- hand/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/hand
-[install]
-install_scripts=$base/lib/hand
