+++ rcj_2023_master/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/rcj_2023_master
+[install]
+install_scripts=$base/lib/rcj_2023_master
